pyproject.toml
[project]
name = "quizbot"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "aiogram (>=3.20.0.post0,<4.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "apscheduler (>=3.11.0,<4.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.12"
mypy = "^1.16.0"
pre-commit = "^4.2.0"

[tool.ruff.lint]
select = ["E", "F", "UP", "B", "SIM", "I"]
ignore = ["F401"]

[tool.mypy]
strict = true
ignore_missing_imports = true
